backend/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
numpy
opencv-python
aiofiles
python-multipart